experiments/7-financial-backtesting/requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
pyarrow>=16.1.0
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.25.0
alpaca-py>=0.31.0
numpy>=1.26.4
nautilus-trader>=1.219.0
plotly>=5.23.0
scipy>=1.13.1
pyyaml>=6.0.2

